add recyclerview version 1

+++ app/src/main/java/com/example/getcellinfos/retrofit/StationInfo.kt
@@ -8,7 +8,7 @@ data class StationInfo(
     var eqp_type: String,
     var enb_id: String,
     var pnu_code: String,
-    var eqp_lat: Double,
-    var eqp_lon: Double,
+    var eqp_lat: Float,
+    var eqp_lon: Float,
     var cell_num: String
 )
